backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
openai
